Fix size_t warning in Visual Studio

+++ src/client/compress_srcml.cpp
@@ -76,7 +76,7 @@ void compress_srcml(const srcml_request_t& /* srcml_request */,
     // write the data into the archive
     std::vector<char> buffer(4092);
     while (true) {
-        ssize_t s = read(*input_sources[0].fd, buffer.data(), buffer.size());
+        ssize_t s = read(*input_sources[0].fd, buffer.data(), (size_t) buffer.size());
         if (s <= 0)
             break;
